backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
faiss-cpu>=1.9.0
sentence-transformers>=2.2.2
numpy>=1.24.3
google-generativeai>=0.3.2
pydantic>=2.5.0
